gl: Use shared surface store
